# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": ""
  },
  "host": "localhost:8080",
  "basePath": "/MUSIC/rest",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/v2/keyspaces/{keyspace}/tables/{tablename}/rows": {
      "get": {
        "tags": [
          "Data Api"
        ],
        "summary": "Select All or Select Specific",
        "description": "",
        "operationId": "select",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "description": "Major Version",
            "required": true,
            "type": "string"
          },
          {
            "name": "X-minorVersion",
            "in": "header",
            "description": "Minor Version",
            "required": false,
            "type": "string"
          },
          {
            "name": "X-patchVersion",
            "in": "header",
            "description": "Patch Version",
            "required": false,
            "type": "string"
          },
          {
            "name": "aid",
            "in": "header",
            "description": "AID",
            "required": true,
            "type": "string"
          },
          {
            "name": "ns",
            "in": "header",
            "description": "Application namespace",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "name": "keyspace",
            "in": "path",
            "description": "Keyspace Name",
            "required": true,
            "type": "string"
          },
          {
            "name": "tablename",
            "in": "path",
            "description": "Table Name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Data Api"
        ],
        "summary": "Insert Into Table",
        "description": "",
        "operationId": "insertIntoTable",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "description": "Major Version",
            "required": true,
            "type": "string"
          },
          {
            "name": "X-minorVersion",
            "in": "header",
            "description": "Minor Version",
            "required": false,
            "type": "string"
          },
          {
            "name": "X-patchVersion",
            "in": "header",
            "description": "Patch Version",
            "required": false,
            "type": "string"
          },
          {
            "name": "aid",
            "in": "header",
            "description": "AID",
            "required": true,
            "type": "string"
          },
          {
            "name": "ns",
            "in": "header",
            "description": "Application namespace",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/JsonTable"
            }
          },
          {
            "name": "keyspace",
            "in": "path",
            "description": "Keyspace Name",
            "required": true,
            "type": "string"
          },
          {
            "name": "tablename",
            "in": "path",
            "description": "Table Name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Data Api"
        ],
        "summary": "Update Table",
        "description": "",
        "operationId": "updateTable",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "description": "Major Version",
            "required": true,
            "type": "string"
          },
          {
            "name": "X-minorVersion",
            "in": "header",
            "description": "Minor Version",
            "required": false,
            "type": "string"
          },
          {
            "name": "X-patchVersion",
            "in": "header",
            "description": "Patch Version",
            "required": false,
            "type": "string"
          },
          {
            "name": "aid",
            "in": "header",
            "description": "AID",
            "required": true,
            "type": "string"
          },
          {
            "name": "ns",
            "in": "header",
            "description": "Application namespace",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/JsonTable"
            }
          },
          {
            "name": "keyspace",
            "in": "path",
            "description": "Keyspace Name",
            "required": true,
            "type": "string"
          },
          {
            "name": "tablename",
            "in": "path",
            "description": "Table Name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Data Api"
        ],
        "summary": "Delete From table",
        "description": "",
        "operationId": "deleteFromTable",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "description": "Major Version",
            "required": true,
            "type": "string"
          },
          {
            "name": "X-minorVersion",
            "in": "header",
            "description": "Minor Version",
            "required": false,
            "type": "string"
          },
          {
            "name": "X-patchVersion",
            "in": "header",
            "description": "Patch Version",
            "required": false,
            "type": "string"
          },
          {
            "name": "aid",
            "in": "header",
            "description": "AID",
            "required": true,
            "type": "string"
          },
          {
            "name": "ns",
            "in": "header",
            "description": "Application namespace",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/JsonTable"
            }
          },
          {
            "name": "keyspace",
            "in": "path",
            "description": "Keyspace Name",
            "required": true,
            "type": "string"
          },
          {
            "name": "tablename",
            "in": "path",
            "description": "Table Name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/v2/keyspaces/{keyspace}/tables/{tablename}": {
      "post": {
        "tags": [
          "Data Api"
        ],
        "summary": "Create Table",
        "description": "",
        "operationId": "createTable",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "description": "Major Version",
            "required": true,
            "type": "string"
          },
          {
            "name": "X-minorVersion",
            "in": "header",
            "description": "Minor Version",
            "required": false,
            "type": "string"
          },
          {
            "name": "X-patchVersion",
            "in": "header",
            "description": "Patch Version",
            "required": false,
            "type": "string"
          },
          {
            "name": "aid",
            "in": "header",
            "description": "AID",
            "required": true,
            "type": "string"
          },
          {
            "name": "ns",
            "in": "header",
            "description": "Application namespace",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/JsonTable"
            }
          },
          {
            "name": "keyspace",
            "in": "path",
            "description": "Keyspace Name",
            "required": true,
            "type": "string"
          },
          {
            "name": "tablename",
            "in": "path",
            "description": "Table Name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Data Api"
        ],
        "summary": "Drop Table",
        "description": "",
        "operationId": "dropTable",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "description": "Major Version",
            "required": true,
            "type": "string"
          },
          {
            "name": "X-minorVersion",
            "in": "header",
            "description": "Minor Version",
            "required": false,
            "type": "string"
          },
          {
            "name": "X-patchVersion",
            "in": "header",
            "description": "Patch Version",
            "required": false,
            "type": "string"
          },
          {
            "name": "aid",
            "in": "header",
            "description": "AID",
            "required": true,
            "type": "string"
          },
          {
            "name": "ns",
            "in": "header",
            "description": "Application namespace",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "name": "keyspace",
            "in": "path",
            "description": "Keyspace Name",
            "required": true,
            "type": "string"
          },
          {
            "name": "tablename",
            "in": "path",
            "description": "Table Name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/{version}/priorityq/keyspaces/{keyspace}/{qname}": {
      "post": {
        "tags": [
          "Q Api"
        ],
        "summary": "Create Q",
        "description": "",
        "operationId": "createQ",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "description": "Major Version",
            "required": true,
            "type": "string"
          },
          {
            "name": "X-minorVersion",
            "in": "header",
            "description": "Minor Version",
            "required": false,
            "type": "string"
          },
          {
            "name": "X-patchVersion",
            "in": "header",
            "description": "Patch Version",
            "required": false,
            "type": "string"
          },
          {
            "name": "aid",
            "in": "header",
            "description": "AID",
            "required": true,
            "type": "string"
          },
          {
            "name": "ns",
            "in": "header",
            "description": "Application namespace",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/JsonTable"
            }
          },
          {
            "name": "keyspace",
            "in": "path",
            "description": "Key Space",
            "required": true,
            "type": "string"
          },
          {
            "name": "qname",
            "in": "path",
            "description": "Table Name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Q Api"
        ],
        "summary": "DropQ",
        "description": "",
        "operationId": "dropQ",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "description": "Major Version",
            "required": true,
            "type": "string"
          },
          {
            "name": "X-minorVersion",
            "in": "header",
            "description": "Minor Version",
            "required": false,
            "type": "string"
          },
          {
            "name": "X-patchVersion",
            "in": "header",
            "description": "Patch Version",
            "required": false,
            "type": "string"
          },
          {
            "name": "aid",
            "in": "header",
            "description": "AID",
            "required": true,
            "type": "string"
          },
          {
            "name": "ns",
            "in": "header",
            "description": "Application namespace",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "name": "keyspace",
            "in": "path",
            "description": "Key Space",
            "required": true,
            "type": "string"
          },
          {
            "name": "qname",
            "in": "path",
            "description": "Table Name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "JsonTable": {
      "type": "object",
      "properties": {
        "columns": {
          "type": "array",
          "description": "Column values",
          "items": {
            "type": "string"
          }
        },
        "consistencyInfo": {
          "type": "object",
          "description": "Consistency level",
          "additionalProperties": {
            "type": "string"
          }
        },
        "conditions": {
          "type": "object",
          "description": "Conditions",
          "additionalProperties": {
            "type": "object"
          }
        },
        "ttl": {
          "type": "string",
          "description": "Time to live information"
        },
        "timestamp": {
          "type": "string",
          "description": "Time stamp"
        }
      },
      "description": "Json model for delete"
    }
  }
}